# docs/apifox/openapi.yaml
openapi: 3.0.3
info:
  title: OptiTree Backend API (Current Implemented Stage)
  version: 0.1.0
  description: |
    This OpenAPI file is generated from the current Go implementation.
    It is intended for direct import into Apifox for debugging.

    Notes:
    - Most business errors are returned with HTTP 200 and `code != 0`.
    - Missing/invalid bearer token may return HTTP 401 from auth middleware.
    - All success envelopes follow: { code, message, data, requestId, timestamp }.
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - BearerAuth: []
...
paths:
...
  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: Register user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              username: demo_user
              email: demo@example.com
              password: pass12345
              agree: true
      responses:
        '200':
          description: Register result
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          user:
                            $ref: '#/components/schemas/User'
              examples:
                success:
                  value:
                    code: 0
                    message: ok
                    data:
                      user:
                        id: user_1741332000000_abcd1234
                        username: demo_user
                        displayName: demo_user
                        email: demo@example.com
                        avatar: null
                        role: user
                        status: active
                        createdAt: '2026-03-08T08:00:00Z'
                        updatedAt: '2026-03-08T08:00:00Z'
                        lastLoginAt: null
                    requestId: e4d8bc8f-3d15-4cb8-8ff4-50bdfcb8d9b1
                    timestamp: '2026-03-08T08:00:00Z'
                invalid:
                  value:
                    code: 40001
                    message: invalid parameter
                    errors: required field missing
                    requestId: e4d8bc8f-3d15-4cb8-8ff4-50bdfcb8d9b1
                    timestamp: '2026-03-08T08:00:00Z'
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Login with username or email
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              username: demo_user
              password: pass12345
              remember: true
      responses:
        '200':
          description: Login result
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LoginData'
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Refresh result
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          accessToken:
                            type: string
                          refreshToken:
                            type: string
                          tokenType:
                            type: string
                            example: Bearer
                          expiresIn:
                            type: integer
                            format: int64
  /api/v1/auth/forgot-password:
    post:
      tags: [Auth]
      summary: Forgot password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Forgot password result
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          message:
                            type: string
  /api/v1/auth/reset-password:
    post:
      tags: [Auth]
      summary: Reset password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, newPassword]
              properties:
                token:
                  type: string
                newPassword:
                  type: string
                  minLength: 8
      responses:
        '200':
          description: Reset password result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      summary: Logout
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Logout result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '401':
          $ref: '#/components/responses/UnauthorizedHttp'
  /api/v1/auth/change-password:
    post:
      tags: [Auth]
      summary: Change password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [oldPassword, newPassword]
              properties:
                oldPassword:
                  type: string
                newPassword:
                  type: string
                  minLength: 8
      responses:
        '200':
          description: Change password result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '401':
          $ref: '#/components/responses/UnauthorizedHttp'
  /api/v1/users/me:
    get:
      tags: [Users]
      summary: Get current user profile
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          user:
                            $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedHttp'
  /api/v1/users/me/update:
    post:
      tags: [Users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                displayName:
                  type: string
                  maxLength: 50
                email:
                  type: string
                  format: email
            example:
              displayName: Demo User
              email: demo@example.com
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          user:
                            $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedHttp'
  /api/v1/users/me/avatar:
    post:
      tags: [Users]
      summary: Upload avatar
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [avatar]
              properties:
                avatar:
                  type: string
                  format: binary
      responses:
        '200':
          description: Upload avatar result
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          avatarUrl:
                            type: string
                            format: uri
        '401':
          $ref: '#/components/responses/UnauthorizedHttp'
  /api/v1/users/me/login-logs:
    get:
      tags: [Users]
      summary: Get login logs
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Login logs
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LoginLogPageData'
        '401':
          $ref: '#/components/responses/UnauthorizedHttp'
...
  /api/v1/projects:
    get:
      tags: [Projects]
      summary: List projects
      parameters:
        - in: query
          name: type
          schema:
            type: string
            enum: [ft, kg]
        - in: query
          name: keyword
          schema:
            type: string
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - in: query
          name: sortBy
          schema:
            type: string
            enum: [created_at, updated_at, name]
            default: created_at
        - in: query
          name: sortOrder
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Project list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ProjectPageData'
    post:
      tags: [Projects]
      summary: Create project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
            example:
              name: Hydraulic analysis
              type: ft
              description: Demo project
              tags: [ai, test]
      responses:
        '200':
          description: Created project
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          project:
                            $ref: '#/components/schemas/Project'
  /api/v1/projects/{projectId}:
    get:
      tags: [Projects]
      summary: Get project by ID
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Project detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          project:
                            $ref: '#/components/schemas/Project'
  /api/v1/projects/{projectId}/update:
    post:
      tags: [Projects]
      summary: Update project
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  maxLength: 100
                description:
                  type: string
                  nullable: true
                tags:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Updated project
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          project:
                            $ref: '#/components/schemas/Project'
  /api/v1/projects/{projectId}/rename:
    post:
      tags: [Projects]
      summary: Rename project (editor role required)
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  maxLength: 100
            example:
              name: Hydraulic analysis renamed
      responses:
        '200':
          description: Renamed project
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          project:
                            $ref: '#/components/schemas/Project'
  /api/v1/projects/{projectId}/delete:
    post:
      tags: [Projects]
      summary: Delete project (admin role required)
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Delete result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
  /api/v1/projects/{projectId}/members:
    get:
      tags: [Members]
      summary: List project members
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Members list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          members:
                            type: array
                            items:
                              $ref: '#/components/schemas/ProjectMember'
  /api/v1/projects/{projectId}/members/invite:
    post:
      tags: [Members]
      summary: Invite project member (admin role required)
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, role]
              properties:
                email:
                  type: string
                  format: email
                role:
                  type: string
                  enum: [viewer, editor, admin]
      responses:
        '200':
          description: Invitation result
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          invitation:
                            $ref: '#/components/schemas/Invitation'
...
  /api/v1/projects/{projectId}/members/{memberId}/remove:
    post:
      tags: [Members]
      summary: Remove member (admin role required)
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/MemberId'
      responses:
        '200':
          description: Remove result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
  /api/v1/projects/{projectId}/versions:
    get:
      tags: [Versions]
      summary: List project versions
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Version list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/VersionPageData'
    post:
      tags: [Versions]
      summary: Create version snapshot
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                label:
                  type: string
                snapshot:
                  type: object
                  additionalProperties: true
            example:
              label: version 2026/3/8 15:00:00
              snapshot:
                nodes: []
                edges: []
      responses:
        '200':
          description: Created version
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          version:
                            $ref: '#/components/schemas/VersionSnapshot'
  /api/v1/projects/{projectId}/versions/{versionId}:
    get:
      tags: [Versions]
      summary: Get version detail
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/VersionId'
      responses:
        '200':
          description: Version detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          version:
                            $ref: '#/components/schemas/VersionSnapshot'
  /api/v1/projects/{projectId}/versions/{versionId}/rollback:
    post:
      tags: [Versions]
      summary: Rollback to version (editor+ required)
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/VersionId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                saveCurrent:
                  type: boolean
                  default: false
      responses:
        '200':
          description: Rollback result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
  /api/v1/projects/{projectId}/versions/{versionId}/delete:
    post:
      tags: [Versions]
      summary: Delete version (admin role required)
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/VersionId'
      responses:
        '200':
          description: Delete version result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    ApiEnvelope:
      type: object
      required: [code, message, requestId, timestamp]
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: ok
        data:
          type: object
          nullable: true
          additionalProperties: true
        requestId:
          type: string
          example: e4d8bc8f-3d15-4cb8-8ff4-50bdfcb8d9b1
        timestamp:
          type: string
          format: date-time
...
    RegisterRequest:
      type: object
      required: [username, email, password, agree]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 20
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        agree:
          type: boolean
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          description: Username or email
        password:
          type: string
        remember:
          type: boolean
          default: false
...